+++ models/face_seg_model.py
@@ -5,10 +5,8 @@
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-import torchvision
 
-from resnet import Resnet18
-# from modules.bn import InPlaceABNSync as BatchNorm2d
+from models.resnet import Resnet18
 
 
 class ConvBNReLU(nn.Module):
